🎨[Refactor] : 프로젝트 구조 변경 및 환경변수 관리를 위한 리팩토링
