--- data/initialFileSystem.tsx
@@ -1,1 +1,0 @@
-// File content removed as initialFileSystem is no longer needed.
