+++ lib/api.ts
@@ -1,11 +1,12 @@
 import axios from 'axios'
 
 const API_KEY = process.env.API_KEY;
+const BASE_SRV_URL = process.env.BASE_SRV_URL;
 
 
 // Axios 인스턴스 생성
 const apiClient = axios.create({
-  baseURL: "http://localhost:8080",
+  baseURL: BASE_SRV_URL + '/files',
   headers: {
     'X-API-KEY': API_KEY, 
     'Content-Type': 'application/json',
@@ -17,7 +18,7 @@ const apiClient = axios.create({
 // 파일 및 폴더 조회 api
 export const fetchFileSystemData = async () => {
   try {
-    const response = await apiClient.get('/api/files')
+    const response = await apiClient.get('')
     return response.data.result
   } catch (error) {
     console.error('Error fetching file system:', error)
@@ -28,7 +29,7 @@ export const fetchFileSystemData = async () => {
 // 파일 또는 폴더 생성 api
 export const createFileOrFolder = async (path: string, type: 'file' | 'folder') => {
   try {
-    const response = await apiClient.post('/api/files/create', null, {
+    const response = await apiClient.post('/create', null, {
       params: { path, type }
     })
     return response.data
@@ -41,7 +42,7 @@ export const createFileOrFolder = async (path: string, type: 'file' | 'folder') 
 // 파일 또는 폴더 이동 api
 export const moveFileOrFolder = async (path: string, to: string) => {
   try {
-    const response = await apiClient.put('/api/files/move', null, {
+    const response = await apiClient.put('/move', null, {
       params: { path, to }
     })
     return response.data
